requirements.txt
torch==2.4.0
torchvision==0.19.0
# python==3.9.*
numpy==1.*


nltk
scikit-learn
networkx
pytest
opencv-python
pytorchcv
requests
pyparsing
zipp
pandas
imgaug
pillow
matplotlib
pytest-shutil
argparse

nbdt
